refactor(orchestrator): slim orchestrator and split livekit/tldraw hooks
Slim down the central orchestrator to improve maintainability and isolate side effects into focused hooks and utils. Event wiring was moved out to utils/* (canvas-creation, selection, navigation/ui-state, mermaid-bridge, pinned-shapes). LiveKit logic was split into small hooks and types (useLkState, useLkToken, useLkAutoConnect, useLkAgentRequest, useLkRoomHandlers, useLkIdentity, utils/lk-types) and exported from the hooks index.

Add a typed tldraw editor bridge (useTldrawEditorBridge) removing the global raw editor, plus a CollaborationLoadingOverlay. Introduce useCanvasEventHandlers(editor, { enabled }) and update LiveKitBus.on to return an unsubscribe (() => void | undefined). Docs updated to clarify the LiveKitBus unsubscribe contract and refactor related tables.

Behavior parity preserved for pinned shapes, mermaid streaming, snapshot broadcaster/receiver and overrides. Quality gates passed (ESLint, TypeScript, Jest). This reduces the orchestrator to ~137 LOC and reorganizes responsibilities for clearer, testable surfaces.
