+++ src/components/ui/livekit/hooks/index.ts
@@ -2,3 +2,9 @@ export * from './types';
 export * from './useAgentDispatch';
 export * from './useLivekitConnection';
 export * from './useRoomEvents';
+export * from './useLkState';
+export * from './useLkToken';
+export * from './useLkAutoConnect';
+export * from './useLkAgentRequest';
+export * from './useLkRoomHandlers';
+export * from './useLkIdentity';
