+++ src/components/ui/livekit/hooks/types.ts
@@ -1,21 +1,7 @@
-export type AgentStatus = 'not-requested' | 'dispatching' | 'joined' | 'failed';
-
-export interface LivekitRoomConnectorState {
-  connectionState: 'disconnected' | 'connecting' | 'connected' | 'error';
-  isMinimized: boolean;
-  participantCount: number;
-  errorMessage: string | null;
-  token: string | null;
-  agentStatus: AgentStatus;
-  agentIdentity: string | null;
-}
-
-export const initialLivekitRoomConnectorState: LivekitRoomConnectorState = {
-  connectionState: 'disconnected',
-  isMinimized: false,
-  participantCount: 0,
-  errorMessage: null,
-  token: null,
-  agentStatus: 'not-requested',
-  agentIdentity: null,
-};
+export {
+  type AgentStatus,
+  type ConnectorConnectionState,
+  type LivekitRoomConnectorAction,
+  type LivekitRoomConnectorState,
+  initialLivekitRoomConnectorState,
+} from './utils/lk-types';
