+++ src/components/ui/tldraw/utils/index.ts
@@ -1,6 +1,11 @@
 export * from './canvas-creation-handlers';
 export * from './canvas-navigation-handlers';
 export * from './canvas-selection-handlers';
+export * from './canvas-pin-handlers';
+export * from './canvas-arrangement-handlers';
+export * from './canvas-list-shapes-handlers';
+export * from './canvas-notes-handlers';
+export * from './canvas-selection-shared';
 export * from './collaborationOverrides';
 export * from './mermaid-bridge';
 export * from './pinned-shapes';
